+++ app/Services/ConfigCacheService.php
@@ -65,6 +65,10 @@ class ConfigCacheService
 				'pixelfed.directory.latest_response',
 				'pixelfed.directory.is_synced',
 				'pixelfed.directory.testimonials',
+
+				'instance.landing.show_directory',
+				'instance.landing.show_explore',
+				'instance.admin.pid',
 				// 'system.user_mode'
 			];