+++ resources/views/admin/settings/home.blade.php
@@ -148,36 +148,43 @@
 	<div class="tab-pane" id="landing" role="tabpanel" aria-labelledby="landing-tab">
 		<div class="form-group mb-0">
 			<div class="ml-n4 mr-n2 p-3 border-top border-bottom">
-				<label class="font-weight-bold text-muted">Name</label>
-				<input class="form-control col-8" name="name" placeholder="Pixelfed" value="{{config_cache('app.name')}}" disabled>
-				<p class="help-text small text-muted mt-3 mb-0">The instance name. Change in Brand section.</p>
+				<p class="mb-0 small">Configure your landing page</p>
 			</div>
 		</div>
 		<div class="form-group mb-0">
 			<div class="ml-n4 mr-n2 p-3 border-bottom">
-				<label class="font-weight-bold text-muted">About</label>
-				<textarea class="form-control" rows="3" name="long_description" disabled>{{config_cache('app.description')}}</textarea>
-				<p class="help-text small text-muted mt-3 mb-0">Description of instance used on about section. Change in Brand section.</p>
-			</div>
-		</div>
-		<div class="form-group mb-0">
-			<div class="ml-n4 mr-n2 p-3 border-bottom">
-				<label class="font-weight-bold text-muted">Other Settings</label>
+				<p class="font-weight-bold text-muted">Discovery</p>
 
 				<div class="my-3">
 					<div class="custom-control custom-checkbox">
-						<input type="checkbox" class="custom-control-input" id="show_directory" name="show_directory" {{ config('instance.landing.show_directory') ? 'checked' : ''}} disabled>
+						<input type="checkbox" class="custom-control-input" id="show_directory" name="show_directory" {{ config_cache('instance.landing.show_directory') ? 'checked' : ''}}>
 						<label class="custom-control-label font-weight-bold" for="show_directory">Show Directory</label>
 					</div>
-					<p class="help-text small text-muted mt-3 mb-0">To disable the Directory, set <code>INSTANCE_LANDING_SHOW_DIRECTORY=false</code> in .env</p>
 				</div>
 
 				<div class="my-3">
 					<div class="custom-control custom-checkbox">
-						<input type="checkbox" class="custom-control-input" id="show_explore_feed" name="show_explore_feed" {{ config('instance.landing.show_explore') ? 'checked' : ''}} disabled>
+						<input type="checkbox" class="custom-control-input" id="show_explore_feed" name="show_explore_feed" {{ config_cache('instance.landing.show_explore') ? 'checked' : ''}}>
 						<label class="custom-control-label font-weight-bold" for="show_explore_feed">Show Explore Feed</label>
 					</div>
-					<p class="help-text small text-muted mt-3 mb-0">To disable the Explore feed, set <code>INSTANCE_LANDING_SHOW_EXPLORE=false</code> in .env</p>
+				</div>
+			</div>
+		</div>
+		<div class="form-group mb-0">
+			<div class="ml-n4 mr-n2 p-3 border-bottom">
+				<p class="font-weight-bold text-muted">Admin Account</p>
+
+				<div class="my-3">
+					<select class="custom-select" name="admin_account_id" style="max-width: 300px;">
+						<option selected disabled>Select an admin account</option>
+						@foreach($availableAdmins as $acct)
+							<option
+								value="{{ $acct->profile_id }}" {!! $currentAdmin && $currentAdmin['id'] == $acct->profile_id ? 'selected' : null !!}
+								>
+								<span class="font-weight-bold">&commat;{{ $acct->username }}</span>
+							</option>
+						@endforeach
+					</select>
 				</div>
 			</div>
 		</div>